+++ packages/frontend/src/App.css.ts
@@ -1,5 +1,5 @@
 import { style } from '@vanilla-extract/css'
-import { spacing, color, rounded, theme } from '@/ui/theme'
+import { dark, spacing, color, rounded, theme, media } from '@/ui/theme'
 import { toggleButton } from '@/ui/components/button/button.css'
 
 export const app = style({
@@ -11,6 +11,11 @@ export const app = style({
   backgroundColor: color.white,
   fontSize: theme.fontSize.base,
   fontFamily: theme.font.sans,
+  ...media({
+    [dark]: {
+      backgroundColor: color.gray[800],
+    },
+  }),
 })
 
 export const header = style({
@@ -20,11 +25,31 @@ export const header = style({
   alignItems: 'center',
   columnGap: spacing[4],
   borderBottom: `1px solid ${color.gray[200]}`,
+  ...media({
+    [dark]: {
+      backgroundColor: color.gray[800],
+      borderColor: color.gray[500],
+    },
+  }),
 })
 
+export const h3 = style({
+  color: color.black,
+  ...media({
+    [dark]: {
+      color: color.gray[50],
+    },
+  }),
+})
 export const select = style({
   width: spacing[8],
   height: spacing[8],
+  color: color.black,
+  ...media({
+    [dark]: {
+      color: color.gray[50],
+    },
+  }),
 })
 export const selectIcon = style({
   width: spacing[5],
@@ -40,6 +65,12 @@ export const optionsButton = style([
     marginLeft: 'auto',
     width: spacing[8],
     height: spacing[8],
+    color: color.black,
+    ...media({
+      [dark]: {
+        color: color.gray[50],
+      },
+    }),
   },
 ])
 export const optionsIcon = style({
@@ -55,6 +86,7 @@ export const optionsPanel = style({
   backgroundColor: color.gray[100],
   border: `1px solid ${color.gray[200]}`,
 })
+
 export const optionsMenu = style({
   display: 'flex',
   flexDirection: 'column',