+++ packages/frontend/src/modules/inspector/inspector.css.ts
@@ -1,4 +1,4 @@
-import { spacing, color } from '@/ui/theme'
+import { dark, spacing, color, media } from '@/ui/theme'
 import { style } from '@vanilla-extract/css'
 
 export const scrollWrapper = style({
@@ -20,6 +20,12 @@ export const h1 = style({
   // TODO: typography
   fontSize: spacing[4],
   fontWeight: 'bold',
+  color: color.black,
+  ...media({
+    [dark]: {
+      color: color.gray[50],
+    },
+  }),
 })
 export const id = style({
   fontSize: spacing[3],
@@ -28,8 +34,13 @@ export const id = style({
   textTransform: 'uppercase',
 })
 export const type = style({
-  color: color.gray[500],
   fontWeight: 400,
+  color: color.gray[500],
+  ...media({
+    [dark]: {
+      color: color.gray[300],
+    },
+  }),
 })
 
 export const content = style({
@@ -41,4 +52,9 @@ export const content = style({
 export const h2 = style({
   color: color.gray[500],
   marginBottom: spacing[1],
+  ...media({
+    [dark]: {
+      color: color.gray[300],
+    },
+  }),
 })