+++ packages/frontend/src/ui/components/splitter/Splitter.css.ts
@@ -1,9 +1,11 @@
 import { style, createVar } from '@vanilla-extract/css'
 import { CSSVarFunction } from '@vanilla-extract/private'
-import { centerChild, color, spacing, transition } from '@/ui/theme'
+import { mobile, centerChild, color, spacing, transition, media } from '@/ui/theme'
 
 const minWidth = spacing[36]
+const minHeight = spacing[12]
 const splitWidth = '1px'
+const splitHeight = '1px'
 
 export const progress: CSSVarFunction = createVar()
 
@@ -18,6 +20,16 @@ export const container = style({
       gridTemplateColumns: `minmax(${minWidth}, ${progress}) ${splitWidth} minmax(${minWidth}, 1fr)`,
     },
   },
+  ...media({
+    [mobile]: {
+      selectors: {
+        '&[data-open="true"]': {
+          gridTemplateColumns: `1fr`,
+          gridTemplateRows: `minmax(${minHeight}, ${progress}) ${splitHeight} minmax(${minHeight}, 1fr)`,
+        },
+      },
+    },
+  }),
 })
 
 export const mainContent = style({
@@ -42,6 +54,11 @@ export const splitHandle = style({
   bottom: 0,
   cursor: 'col-resize',
   userSelect: 'none',
+  ...media({
+    [mobile]: {
+      cursor: 'row-resize',
+    },
+  }),
 })
 
 export const toggle = style({
@@ -59,6 +76,13 @@ export const toggle = style({
   ':hover': {
     backgroundColor: color.gray[500],
   },
+  ...media({
+    [mobile]: {
+      width: spacing[6],
+      height: spacing[6],
+      right: spacing[3],
+    },
+  }),
 })
 export const x = style({
   width: spacing[4],