+++ packages/frontend/src/ui/theme/utils.ts
@@ -1,4 +1,5 @@
-import { Many } from '@solid-primitives/utils'
+import { asArray, Many } from '@solid-primitives/utils'
+import { StyleRule } from '@vanilla-extract/css'
 import type { CSSPropertiesWithVars } from '@vanilla-extract/css/dist/declarations/src/types'
 import { clsx } from 'clsx'
 import { Property } from 'csstype'
@@ -68,3 +69,26 @@ export const transition = (
   transitionDelay: delay,
   transitionTimingFunction: easing,
 })
+
+export const media = (
+  rules: Many<({ rule: Many<string> } & StyleRule) | Record<string, StyleRule>>,
+): {
+  '@media': Record<string, StyleRule>
+} => {
+  const media: Record<string, StyleRule> = {}
+  for (const obj of asArray(rules)) {
+    if ('rule' in obj) {
+      const { rule, ...styles } = obj
+      const calcRule = ['screen', ...asArray(rule)].join(' and ')
+      media[calcRule] = styles
+    } else {
+      for (const [rule, styles] of Object.entries(obj)) {
+        media[`screen and ${rule}`] = styles
+      }
+    }
+  }
+  return { '@media': media }
+}
+export const dark = '(prefers-color-scheme: dark)'
+export const mobile = '(max-width: 640px)'
+export const touch = '(hover: none)'