+++ packages/extension/CHANGELOG.md
@@ -1,5 +1,19 @@
 # solid-devtools-extension
 
+## 0.15.3
+
+### Patch Changes
+
+- Updated dependencies [aa9f0bb]
+  - @solid-devtools/frontend@0.0.5
+
+## 0.15.2
+
+### Patch Changes
+
+- Updated dependencies [95aaad9]
+  - @solid-devtools/frontend@0.0.4
+
 ## 0.15.1
 
 ### Patch Changes